rest-api: consolidate Operating System tests into one file per type
Fold operatingsystem_ownership_test.go and operatingsystem_visibility_test.go
into operatingsystem_test.go (handler), and
operatingsystem_providervisibility_test.go into operatingsystem_test.go (db
model), following the one-test-file-per-type convention. No test logic changes.
